% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sobolSequence.R
\name{sobolSequence.dimF2MinMax}
\alias{sobolSequence.dimF2MinMax}
\title{get minimum and maximum F2 dimension number.}
\usage{
sobolSequence.dimF2MinMax(dimR)
}
\arguments{
\item{dimR}{dimention.}
}
\value{
supportd minimum and maximum F2 dimension number for specified digitalNet.
}
\description{
get minimum and maximum F2 dimension number.
}
