% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sobolSequence.R
\name{sobolSequence.dimMinMax}
\alias{sobolSequence.dimMinMax}
\title{get minimum and maximum dimension number of Sobol Sequence}
\usage{
sobolSequence.dimMinMax()
}
\value{
supportd minimum and maximum dimension number.
}
\description{
get minimum and maximum dimension number of Sobol Sequence
}
