% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sobolSequence.R
\name{sobolSequence.maxCount}
\alias{sobolSequence.maxCount}
\title{get maximum number of points.}
\usage{
sobolSequence.maxCount(dim)
}
\arguments{
\item{dim}{dimension.}
}
\value{
supported maximum number of points.
}
\description{
get maximum number of points.
}
