% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sobolSequence.R
\name{sobolSequence.points}
\alias{sobolSequence.points}
\title{get points from SobolSequence}
\usage{
sobolSequence.points(dimR, dimF2 = 10, count, digitalShift = FALSE)
}
\arguments{
\item{dimR}{dimention.}

\item{dimF2}{F2-dimention of each element.}

\item{count}{number of points.}

\item{digitalShift}{use digital shift or not.}
}
\value{
matrix of points where every row contains dimR dimensional point.
}
\description{
get points from SobolSequence
}
