% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sobolSequence.R
\name{sobolSequence.maxCount}
\alias{sobolSequence.maxCount}
\title{get minimum and maximum F2 dimension number.}
\usage{
sobolSequence.maxCount(dim)
}
\arguments{
\item{dim}{dimension.}
}
\value{
supported maximum number of points.
}
\description{
get minimum and maximum F2 dimension number.
}
