% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sobolSequence.R
\name{sobolSequence.points}
\alias{sobolSequence.points}
\title{get points from SobolSequence}
\usage{
sobolSequence.points(dim, count, digitalShift = FALSE)
}
\arguments{
\item{dim}{dimension.}

\item{count}{number of points.}

\item{digitalShift}{use digital shift or not.}
}
\value{
matrix of points where every row contains dimR dimensional point.
}
\description{
This R version does not returns coordinate value zero,
but returns value very near to zero, 2^-64.
}
